Feature update Modified the service and controller to pass test case
