+++ quora-service/src/main/java/com/upgrad/quora/service/business/SignInBusinessService.java
@@ -4,7 +4,6 @@ import com.upgrad.quora.service.dao.UserDao;
 import com.upgrad.quora.service.entity.UserAuthEntity;
 import com.upgrad.quora.service.entity.UserEntity;
 import com.upgrad.quora.service.exception.AuthenticationFailedException;
-import com.upgrad.quora.service.exception.SignUpRestrictedException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -12,11 +11,10 @@ import javax.transaction.Transactional;
 import java.time.ZonedDateTime;
 
 /*
-This class implements the user signup feature
+This class implements the user sign-in feature
  */
 @Service
-public class UserBusinessService {
-
+public class SignInBusinessService {
 
     //Creating an instance to encrypt the password
     @Autowired
@@ -25,43 +23,6 @@ public class UserBusinessService {
     //Creating an instance to access DB
     @Autowired
     private UserDao userDao;
-
-    /*
-    This method will get the user as input and performs the task of creating the user.
-    @Parm UserEntity
-    @Return UserEntity
-     */
-    @Transactional
-    public UserEntity performSignUp(UserEntity newUser) throws SignUpRestrictedException {
-        if (userDao.getUserByUserName(newUser.getUserName()) != null) {
-            throw new SignUpRestrictedException("SGR-001", "Try any other Username, this Username has already been taken");
-        } else if (userDao.getUserByEmail(newUser.getEmail()) != null) {
-            throw new SignUpRestrictedException("SGR-002", "This user has already been registered, try with any other emailId");
-        } else {
-            return userDao.RegisterUser(newUser);
-        }
-
-    }
-
-    /*This method will signout the user
-    @Parm - accessToken
-     */
-
-    @Transactional
-    public UserAuthEntity validateAccessToken(String accessToken) throws SignUpRestrictedException {
-        UserAuthEntity userAuthEntity = userDao.getUserAuthEntityByAccessToken(accessToken);
-        if (userAuthEntity == null) {
-            throw new SignUpRestrictedException("SGR-001", "User is not Signed in");
-        } else {
-            return userAuthEntity;
-        }
-    }
-
-    @Transactional
-    public void performSignOut(UserAuthEntity userAuthEntity) {
-        userDao.updateUserAuthEntity(userAuthEntity);
-    }
-
     /*
     @parm userName,passWord
     @Return UserAuthEntity
@@ -69,7 +30,7 @@ public class UserBusinessService {
     respective task.
      */
     @Transactional
-    public UserAuthEntity PerformUserSignin(String userName, String passWord) throws AuthenticationFailedException {
+    public UserAuthEntity PerformUserSignIn(String userName, String passWord) throws AuthenticationFailedException {
         UserEntity user = userDao.getUserByUserName(userName);
         if (user == null) {
             throw new AuthenticationFailedException("ATH-001", "This username does not exist");
@@ -94,6 +55,4 @@ public class UserBusinessService {
 
 
     }
-
 }
-
